fix errors

+++ dataset/bad_randomness/guess_the_random_number.sol
@@ -12,7 +12,7 @@ contract GuessTheRandomNumberChallenge {
     constructor()  payable  {
         require(msg.value == 1 ether);
         
-        answer = uint8(keccak256(abi.encodePacked(blockhash(block.number - 1), block.timestamp)));
+        answer = uint8(uint256(keccak256(abi.encodePacked(blockhash(block.number - 1), block.timestamp))));
     }
 
     function isComplete() public view returns (bool) {
